requirements.txt
kafka-python
pyspark==3.1.2
apache-airflow-providers-apache-spark
apache-airflow-providers-apache-kafka
confluent-kafka